Struct multi_semaphore::Semaphore

pub struct Semaphore { /* fields omitted */ }

A counting, blocking, semaphore.

Semaphores are a form of atomic counter where access is only granted if the counter is a positive value. Each acquisition will block the calling thread until the counter is positive, and each release will increment the counter and unblock any threads if necessary. This library allows getting a count greater than 1. If a thread acquires 4 resources, the thread will block until the counter is 4 or greater. Each release will increment the counter and unblock any threads if necessary.

Examples

use multi_semaphore::Semaphore;

// Create a semaphore that represents 6 resources
let sem = Semaphore::new(6);

// Acquire one or more of the resources
sem.acquire();
sem.acquire_many(2);

// Acquire one or more of the resources for a limited period of time
{
    let _guard = sem.access();
    // ...
} // resource is released here
{
    let _guard = sem.access_many(3);
    // ...
} // resources are released here

// Release our initially acquired resource
sem.release();
// Take care to relase the number of resources you intend, rather than too few or too many.
// Using guards from `access_many(n)` is preferred.
sem.release_many(2);

Implementations

impl Semaphore

pub fn new(count: isize) -> Semaphore

Creates a new semaphore with the initial count specified.

The count specified can be thought of as a number of resources, and a call to acquire or access will block until at least one resource is available. It is valid to initialize a semaphore with a negative count.

pub fn acquire(&self)

Acquires a resource of this semaphore, blocking the current thread until it can do so.

This method will block until the internal count of the semaphore is at least 1.

pub fn acquire_many(&self, amount: isize)

Acquires one or more resources of this semaphore, blocking the current thread until it can do so.

This method will block until the internal count of the semaphore is at least amount.

pub fn release(&self)

Release a resource from this semaphore.

This will increment the number of resources in this semaphore by 1 and will notify any pending waiters in acquire or access if necessary.

pub fn release_many(&self, amount: isize)

Release one or more resources from this semaphore.

This will increment the number of resources in this semaphore by 1 and will notify any pending waiters in acquire or access if necessary.

pub fn access(&self) -> SemaphoreGuard

Acquires a resource of this semaphore, returning an RAII guard to release the semaphore when dropped.

This function is semantically equivalent to an acquire followed by a release when the guard returned is dropped.

pub fn access_many(&self, amount: isize) -> SemaphoreGuard

Acquires one or more resources of this semaphore, returning an RAII guard to release the semaphore when dropped.

This function is semantically equivalent to an acquire_many(n) followed by a release_many(n) when the guard returned is dropped.